JRDMessagingConfiguration moved to config package
