+++ src/test/java/uk/gov/hmcts/reform/orgrolemapping/servicebus/JRDTopicConsumerTest.java
@@ -9,6 +9,7 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.testcontainers.shaded.com.fasterxml.jackson.core.JsonProcessingException;
 import org.testcontainers.shaded.com.fasterxml.jackson.databind.ObjectMapper;
+import uk.gov.hmcts.reform.orgrolemapping.config.JRDMessagingConfiguration;
 import uk.gov.hmcts.reform.orgrolemapping.controller.advice.exception.InvalidRequest;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.UserRequest;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.enums.UserType;
@@ -35,12 +36,15 @@ class JRDTopicConsumerTest {
     @Mock
     public SubscriptionClient subscriptionClient;
 
+    @Mock
+    JRDMessagingConfiguration configuration;
+
     JRDTopicConsumer sut;
 
     @BeforeEach
     public void setUp() {
         MockitoAnnotations.openMocks(this);
-        sut = new JRDTopicConsumer(bulkAssignmentOrchestrator, deserializer);
+        sut = new JRDTopicConsumer(bulkAssignmentOrchestrator, deserializer, configuration);
     }
 
     @Test
